+++ src/vertex.h
@@ -125,9 +125,9 @@ class vertex
         int ruleLikes()const{return RuleLikes;}
 #endif
 #if SMALLMEMORY
-        int nlemmatise(trainingPair * pairs,int n,bool InputRight,ptrdiff_t skip);
+        int nlemmatise(trainingPair * pairs,int n,bool InputRight,int skip);
 #else
-        int nlemmatise(trainingPair * pairs,bool InputRight,ptrdiff_t skip);
+        int nlemmatise(trainingPair * pairs,bool InputRight,int skip);
 #endif
 #if AMBIGUOUS
         void markAmbiguousForNextRound(trainingPair * pair);
